refactor: Improve `*::Error` variants, and modules
*Better `Error`s*

The `Error` data constructors defined by me, weren't that descriptive,
and so the error messages in the client were impossible to deal with due
to how vague they were. If something bad happens, the client has no idea
if it was the server's fault, and even if it was there fault, they have
no idea what they did wrong. So now there's enough context to
distinguish if it's a 4xx or 5xx, and if it's the former, what can be
done to fix it.

*More organized modules*

The modules weren't as organized before so that is _slightly_ better now.
The DB connection stuff remain in the `db.rs` module, but the functions
that queried the db for data now live in a `url.rs` module.

There are still more things to be refactored. This just takes care of a
sizable chunk of them.